backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pydantic-settings==2.3.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
google-api-python-client==2.127.0
google-auth==2.32.0
watchfiles==0.22.0
bcrypt==4.1.2
